replace old cli tool
