+++ libtiledbvcf/cmake/Superbuild.cmake
@@ -65,7 +65,6 @@ set(INHERITED_CMAKE_ARGS
 
 # These includes modify the EXTERNAL_PROJECTS variable.
 
-include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/Modules/FindClipp_EP.cmake)
 include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/Modules/FindCLI11_EP.cmake)
 include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/Modules/FindHTSlib_EP.cmake)
 include(${CMAKE_CURRENT_SOURCE_DIR}/cmake/Modules/FindTileDB_EP.cmake)
